requirements.txt
langchain_openai
langchain-community
streamlit
pandas
altair
plotly
langchain
mlflow==2.12.1
tiktoken
openai
python-dotenv
langchain_core
langchain-google-genai
pillow
langchain-experimental
matplotlib